getting more stats from benchmarks (#3142)
As far as I understand, the consensus (I reached with Sergey) is that this is to be merged as is (which is 'made usable, but the codebase needs further cleanup'), as having some common base should make further development and benchmarking easier.

If there are some comments, feel free to make them; I still plan one more iteration of cleanup (mostly for documentation, decreasing amount of code duplication, and addressing the comments).

Co-authored-by: Sergey Kulik <104143901+zxqfd555-pw@users.noreply.github.com>
GitOrigin-RevId: 61d19a89815123f30771111965c8d36112550cfc