Krzysiek/flink benchmarks (#3248)
Flink / scala implementation of the wordcount benchmark.

As far as I understand, the consensus is that this should be merged, even though nobody read-approved it. Current version supports single-worker, single-core, but can be easily extended to the multi-worker, multi-core setting.

Changes to places that should not be changed (docker compose for kstreams and variables_ext.env) are made by a formatter that does not allow to keep empty line at the end of the file.

GitOrigin-RevId: 9df0105036f357576e862ba07b33b1c9f7ec8ec0