add support for defining CPU pools for benchmarks (#3355)
Support for CPU pools (mostly for making Flink use only single core; but perhaps also for other things).

Remark: now flink needs some adjustment to be compared in a fair way, i.e. we need a variant with micro-batching (now it answers for each line, if the output size with batching is small, it can't be competitive) and disable internal parallelism entirely (if we limit it to 1 cpu, we also need limit number of threads it uses; otherwise it won't be a fair comparison, as it looses on switching between them)
GitOrigin-RevId: d02bfce0ead60df95844eaf507579c11f8ca7899