better minibatched flink (no stream API) (#3482)
Merge flink wordcount, using Table (without stream API) to develop. Implemented for the purpose of having minibatched version of Flink code that has higher throughput.
The matter of local env (whether setting anything there changes anything if we use TalbeEnvironment) is addressed in the branch with setting for parallel runs (as overthere it is more relevant).

GitOrigin-RevId: 67e3574bcf675597727b205a7c76158182b17f4a